Add files via upload
This individual project was completed in accordance with the Introduction to Cyber Security (IE2022) for the first semester of the second year of the Cyber Security Specialization program. 
 
This involves researching and exploring methods for gaining unauthorized access to drones and manipulating their functions. The project will use various technologies such as wireless communication protocols and penetration testing techniques to achieve this. The goal of the project is to understand the security vulnerabilities of drones and identify ways to improve their protection. Additionally, the project aims to investigate potential future uses of drone hacking such as in the field of cybersecurity and for defensive security measures.